requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
django-filter==24.1
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34 